New argument validation algorithm & vararg to array transformation
Add: A new clearer algorithm for validating function arguments against the defined parameters.
Add: Vararg arguments are transformed to arrays (for example, the argument array ["testString", 0, 1, 2] would be transformed to ["testString, [0, 1, 2]]). The type of the inner vararg array is equal to the type of the parameter that declares the array/vararg.